# api/openapi.yaml
openapi: 3.0.3
info:
  title: OrbitJob API
  version: 0.1.0
  description: Code-first OpenAPI document generated from OrbitJob HTTP route metadata and DTOs.
paths:
...
  /api/v1/functions:
    get:
      operationId: listFunctions
      summary: List function definitions
      description: List the function definitions in the tenant, newest first. A function is a tenant-owned definition row; a key scoped to a resource group sees only that group's functions. limit defaults to 50 when omitted.
      tags:
      - Functions
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: tenant_id
        in: query
        schema:
          type: string
          minLength: 26
          maxLength: 26
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
          minimum: 1
          maximum: 100
      - name: offset
        in: query
        schema:
          type: integer
          default: 0
          minimum: 0
      responses:
        "200":
          description: Function list
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionListResponse"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/functions/{id}:
    get:
      operationId: getFunction
      summary: Get one function definition
      description: "Get one function definition by id. A function has no schedule: it is invoke-only, and every execution-shaping field is pinned into the revision each invocation runs."
      tags:
      - Functions
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: tenant_id
        in: query
        schema:
          type: string
          minLength: 26
          maxLength: 26
      responses:
        "200":
          description: Function detail
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionItem"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Function not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/functions/{id}/invoke:
    post:
      operationId: invokeFunction
      summary: Invoke a function
      description: "Invoke a function by publishing a JobRun custom resource, exactly as a manual job trigger does; the operator materializes the ledger run from the pinned revision. The request carries no body: v1 functions are parameterless, and parameters are baked into the definition. With wait_seconds (capped at 60) the call polls the custom resource until a terminal phase or the budget expires, then answers with the phase observed. Every invocation is a cold pod, so latency is honest seconds, not milliseconds -- and the only result channel is the exit outcome: the platform returns a phase and a duration, never logs or a response body. Invoking a paused function is a conflict. Use the X-OrbitJob-Idempotency-Key header to prevent duplicate invocations."
      tags:
      - Functions
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: wait_seconds
        in: query
        schema:
          type: integer
          minimum: 0
          maximum: 60
      - name: X-OrbitJob-Idempotency-Key
        in: header
        schema:
          type: string
          maxLength: 128
      responses:
        "200":
          description: Existing invocation returned for idempotent replay, or the wait expired and the async reference is repeated
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionInvokeResult"
        "201":
          description: Invocation published
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionInvokeResult"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Function not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The function is paused, or has no active revision yet
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/functions/{id}/runs:
    get:
      operationId: listFunctionRuns
      summary: List function invocations
      description: "List a function's terminal invocations from the function_runs read model, newest first. The model records outcomes only -- success, failed, canceled -- because live progress lives on the run's JobRun custom resource and the ledger row."
      tags:
      - Functions
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: tenant_id
        in: query
        schema:
          type: string
          minLength: 26
          maxLength: 26
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
          minimum: 1
          maximum: 100
      responses:
        "200":
          description: Function run list
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionRunListResponse"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Function not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/functions/{id}/runs/{run_id}:
    get:
      operationId: getFunctionRun
      summary: Get one function invocation
      description: Get one terminal invocation by its deterministic run id, which is derived from the ledger occurrence key so a replayed recording addresses the same row.
      tags:
      - Functions
      parameters:
      - name: X-Trace-ID
        in: header
        schema:
          type: string
          maxLength: 128
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          maxLength: 64
      responses:
        "200":
          description: Function run detail
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionRunItem"
        "400":
          description: Invalid request
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Function or run not found
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Minimum seconds until the request may be retried (default 1)
              schema:
                type: string
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          headers:
            X-Trace-ID:
              description: Trace identifier echoed back to clients. Generated when request does not provide one.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    APIError:
      type: object
      properties:
        code:
          type: string
          enum:
          - MALFORMED_REQUEST
          - VALIDATION_ERROR
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - RATE_LIMITED
          - INTERNAL_ERROR
          - SERVICE_UNAVAILABLE
        field:
          type: string
        message:
          type: string
      required:
      - code
      - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/APIError"
      required:
      - error
    FunctionInvokeResult:
      type: object
      properties:
        created:
          type: boolean
        name:
          type: string
        namespace:
          type: string
        occurrence_key:
          type: string
        phase:
          type: string
        trigger:
          type: string
      required:
      - namespace
      - name
      - occurrence_key
      - trigger
      - phase
      - created
    FunctionItem:
      type: object
      properties:
        args:
          type: array
          items:
            type: string
        command:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        description:
          type: string
        history_failed:
          type: integer
        history_success:
          type: integer
        id:
          type: integer
          format: int64
        image:
          type: string
        labels:
          type: object
          additionalProperties: true
        name:
          type: string
        resource_group_id:
          type: string
        retry_limit:
          type: integer
        status:
          type: string
        tenant_id:
          type: string
        timeout_seconds:
          type: integer
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
      required:
      - id
      - tenant_id
      - resource_group_id
      - name
      - description
      - status
      - image
      - command
      - args
      - timeout_seconds
      - retry_limit
      - history_success
      - history_failed
      - labels
      - version
      - created_at
      - updated_at
    FunctionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FunctionItem"
      required:
      - items
    FunctionRunItem:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        function_id:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        run_id:
          type: string
        started_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        triggered_at:
          type: string
          format: date-time
        version:
          type: integer
      required:
      - id
      - run_id
      - function_id
      - status
      - triggered_at
      - version
      - created_at
    FunctionRunListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FunctionRunItem"
      required:
      - items
...